Made EditEventTypesActivity use the new EditEventTypeLineView view
* Changed EventType::_id be long rather than int

Still to do before this part is finished:

* Test & debug the save & cancel buttons
* Make the delete button do something
* Make the delete button a trashcan!
* Toast notification when saving

+++ src/org/belligero/nautilus/life/logger/LifeLoggerActivity.java
@@ -112,17 +112,25 @@ public class LifeLoggerActivity extends Activity {
 		if (LifeLoggerActivity.instance == null) LifeLoggerActivity.instance = this;
 		
 		_logButtons.removeAllViews();
-		HashMap<Integer, String> idNameMapping = new HashMap<Integer, String>(); // TODO Get rid of this hashmap
+		HashMap<Long, String> idNameMapping = new HashMap<Long, String>(); // TODO Get rid of this hashmap
 
 		Calendar cal = Calendar.getInstance();
 		
 		// Load the buttons
 		EventTypeIterator eventTypeIterator = _dbHelper.eventTypeHandler.fetchAllEventTypes();
 		for (EventType eventType : eventTypeIterator) {
-			idNameMapping.put(Integer.valueOf(eventType.getID()), eventType.getName());
+			idNameMapping.put(
+					eventType.getID(),
+					eventType.getName()
+				);
 			
 			if (eventType.isActive()) {
-				_logButtons.addView(new LogEventLineView(this, eventType));
+				_logButtons.addView(
+						new LogEventLineView(
+								this,
+								eventType
+							)
+					);
 			}
 		}
